+++ plutus-ledger/src/Ledger/Address.hs
@@ -31,7 +31,6 @@ module Ledger.Address
     xprvToStakingCredential,
     xprvToStakePubKey,
     xprvToStakePubKeyHash,
-    toCardanoAddressInConway,
   )
 where
 
@@ -50,7 +49,6 @@ import Ledger.Scripts
   ( ScriptHash (..),
     StakeValidatorHash (..),
     ValidatorHash (..),
-    toCardanoAddressInConway,
   )
 import PlutusLedgerApi.V1.Address as Export hiding (pubKeyHashAddress, toScriptHash)
 import PlutusLedgerApi.V1.Credential