+++ plutus-script-utils/src/Plutus/Script/Utils/V1/Typed/Scripts/Validators.hs
@@ -31,7 +31,9 @@ import Plutus.Script.Utils.Scripts
   ( Language (PlutusV1),
     Validator,
     Versioned (Versioned),
+    toMintingPolicyHash,
     toValidator,
+    toValidatorHash,
   )
 import Plutus.Script.Utils.Typed
   ( DatumType,
@@ -45,7 +47,7 @@ import Plutus.Script.Utils.Typed
     mkUntypedValidator,
     vForwardingMintingPolicy,
   )
-import Plutus.Script.Utils.V1.Scripts qualified as Scripts
+import Plutus.Script.Utils.V1.Scripts ()
 import Plutus.Script.Utils.V1.Typed.Scripts.MonetaryPolicies qualified as MPS
 import PlutusCore.Default (DefaultUni)
 import PlutusCore.Version (plcVersion100)
@@ -62,10 +64,10 @@ validatorToTypedValidator val =
     { tvValidator = Versioned val PlutusV1,
       tvValidatorHash = hsh,
       tvForwardingMPS = Versioned mps PlutusV1,
-      tvForwardingMPSHash = Scripts.mintingPolicyHash mps
+      tvForwardingMPSHash = toMintingPolicyHash mps
     }
   where
-    hsh = Scripts.validatorHash val
+    hsh = toValidatorHash val
     mps = MPS.mkForwardingMintingPolicy hsh
 
 -- | Make a 'TypedValidator' from the 'CompiledCode' of a validator script and its wrapper.